Use montagu-vault-py
