+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "backup/aws/vault"]
+	path = backup/aws/vault
+	url = https://github.com/vimc/montagu-vault-py
